refactor : Refactor int 형에서 BigInteger로 변경

+++ src/main/java/com/example/practice/product/domain/Product.java
@@ -51,7 +51,7 @@ public class Product {
 		return image;
 	}
 
-	public int getMoney() {
+	public BigInteger getMoney() {
 		return money.getPrice();
 	}
